requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
construct==2.5.3
Django==4.2.7
django-elasticsearch-dsl==8.0
# django-haystack==3.2.1
djangorestframework==3.14.0
# drf-haystack==1.8.13
elastic-transport==8.12.0
elasticsearch>=7.0.0
elasticsearch-dsl==8.11.0
idna==3.6
pefile==2023.2.7
python-dateutil==2.8.2
python-ptrace==0.9.8
pytz==2023.4
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18